# Conflicts: #	LedDriver.cpp #	LedDriver.h #	Letters.h #	Modes.h #	Qlockwork.ino #	Renderer.cpp #	Renderer.h #	src/Configuration.h #	src/Events.h #	src/Readme.txt #	src/Settings.cpp
