+++ platformio.ini
@@ -1,0 +1,28 @@
+; PlatformIO Project Configuration File
+;
+;   Build options: build flags, source filter
+;   Upload options: custom upload port, speed and extra flags
+;   Library options: dependencies, extra library storages
+;   Advanced options: extra scripting
+;
+; Please visit documentation for the other options and examples
+; https://docs.platformio.org/page/projectconf.html
+
+[env:esp12e]
+platform = espressif8266@~3.0.0
+board = esp12e
+framework = arduino
+lib_deps = 
+	arduino-libraries/ArduinoHttpClient@^0.4.0
+	adafruit/Adafruit Unified Sensor@^1.1.4
+	crankyoldgit/IRremoteESP8266@^2.7.10
+	paulstoffregen/Time@^1.6
+	arduino-libraries/Arduino_JSON @ ^0.2.0
+	mysensors/MySensors NodeManager @ ^1.8.0
+	jchristensen/DS3232RTC @ ^2.0.1
+	adafruit/Adafruit NeoPixel @ ^1.12.0
+	adafruit/DHT sensor library @ ^1.4.6
+	adafruit/Adafruit Unified Sensor @ ^1.1.14
+	bblanchon/ArduinoJson@^7.1.0
+monitor_speed = 115200
+upload_port = COM6
