+++ src/misc/BOM.txt
@@ -1,0 +1,30 @@
+--------------------------------------------------------------
+
+1x        WeMos D1 mini
+115x      SK6812 or WS2812B RGB(W) LED-Stripe, 30 LEDs per meter
+1x        DS3231 RTC Module
+1x        DHT22 Module
+1x 145475 LDR A9013
+1x        HX1838 IR-Remote
+1x 171085 IR-Receiver TSOP31240
+3x        Button
+1x        Powersupply 5V, 4A
+
+--------------------------------------------------------------
+
+1x Baseplate 450 x 450 x 2 mm
+
+Outer frame:
+2x Bar 450 x 20 x 15 mm
+2x Bar 420 x 20 x 15 mm
+
+Inner frame:
+2x Bar 355 x 20 x 15 mm
+2x Bar 325 x 20 x 15 mm
+8x Supermagnets 10 x 5 mm
+
+Light grid:
+12x Bar 325 x 20 x 5 mm
+99x Bar  23 x 20 x 5 mm
+
+--------------------------------------------------------------
